Overview now checks github api for recent commit
overview.py file now checks github api for the most recent commit on my github account. For the future, it should check the most recent commit in the last 2 days but currently the app has a statis date set.

Additionally, the blog and commit fetching methods were moved to a separate file.